refactor: CartItem의 로직을 useCartItem 커스텀 훅으로 분리
