+++ src/constants/cartConfig.ts
@@ -1,3 +1,5 @@
 export const SHIPPING_FEE = 3_000;
 
 export const SHIPPING_FEE_THRESHOLD = 100_000;
+
+export const MINIMUM_CART_ITEM = 1;
